#2 Allow Xamarin to pull back a list of calendars for both iOS and Android
GetCalendarsAsync:
- Check/Request calendar read permissions
- Declare calendar instance/singleton
- Request List of iOS/Android platform specific calenders
- Cast to new shared List<DeviceCalendar> which obtains values for
  * Id - Platform specific android generally holds number ids whilst iOS uses guids
  * Name - Name of each calendar
  * IsReadOnly - Can this calendar be edited
- Currently the sample app only displays the name attribute, but we could indicate read only access and ids at a later state.
