--- Xamarin.Essentials/Types/CalendarExtensions.android.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Xamarin.Essentials
-{
-    // Android
-    public partial class DeviceSpecificAttendeeDetails
-    {
-        public bool IsAttending { get; set; }
-
-        public AttendanceStatus Status { get; set; }
-    }
-
-    public enum AttendanceStatus
-    {
-        None,
-        Accepted,
-        Declined,
-        Invited,
-        Tentative
-    }
-}
