+++ Xamarin.Essentials/Types/DeviceCalendar.shared.cs
@@ -4,18 +4,31 @@ using System.Collections.Generic;
 namespace Xamarin.Essentials
 {
     [Preserve(AllMembers = true)]
-    public class DeviceCalendar
+    public class DeviceCalendar : ICalendar
     {
         public string Id { get; set; }
 
         public string Name { get; set; }
+
+        public bool IsReadOnly { get; set; }
+
+        // public List<Event> EventList { get; set; }
     }
 
-    [Preserve(AllMembers = true)]
-    public class DeviceEvent
+    public interface ICalendar
     {
-        public string Id { get; set; }
+        string Id { get; set; }
 
+        string Name { get; set; }
+
+        bool IsReadOnly { get; set; }
+    }
+
+    public class Event
+    {
+        public int Id { get; set; }
+
+        // Calendar Id this event is for
         public string CalendarId { get; set; }
 
         public string Title { get; set; }
@@ -24,50 +37,78 @@ namespace Xamarin.Essentials
 
         public string Location { get; set; }
 
-        public bool AllDay
-        {
-            get => !EndDate.HasValue;
-            set
-            {
-                if (value)
-                {
-                    EndDate = null;
-                }
-                else
-                {
-                    EndDate = StartDate;
-                }
-            }
-        }
+        public bool AllDay { get; set; }
 
-        public DateTimeOffset StartDate { get; set; }
+        public DateTime Start { get; set; }
 
-        public TimeSpan? Duration
-        {
-            get => EndDate.HasValue ? EndDate - StartDate : null;
-            set
-            {
-                if (value.HasValue)
-                {
-                    EndDate = StartDate.Add(value.Value);
-                }
-                else
-                {
-                    EndDate = null;
-                }
-            }
-        }
+        public DateTime End { get; set; }
 
-        public DateTimeOffset? EndDate { get; set; }
+        public bool HasRecurrence { get; set; }
 
-        public IEnumerable<DeviceEventAttendee> Attendees { get; set; }
+        public RecurrenceRule RecurrenceFrequency { get; set; }
+
+        public bool HasReminder { get; set; }
+
+        public List<Reminder> Reminders { get; set; }
     }
 
-    [Preserve(AllMembers = true)]
-    public class DeviceEventAttendee
+    public class Reminder
+    {
+        public static int Id { get; set; }
+
+        public int MinutesPriorToEvent { get; set; }
+    }
+
+    public class RecurrenceRule
+    {
+        public int TotalOccurences { get; set; }
+
+        public int Interval { get; set; }
+
+        public DateTime EndDate { get; set; }
+
+        public RecurrenceFrequency Frequency { get; set; }
+
+        // Only allow event to occur on these days [not available for daily]
+        public List<DayOfTheWeek> DaysOfTheWeek { get; set; }
+
+        public List<int> DaysOfTheMonth { get; set; }
+
+        public List<int> WeeksOfTheYear { get; set; }
+
+        public List<int> SetPositions { get; set; }
+    }
+
+    public class Attendee
     {
         public string Name { get; set; }
 
         public string Email { get; set; }
+
+        // public IAttendeeDetails AttendeeDetails => DeviceSpecificAttendeeDetails;
+    }
+
+    public interface IAttendeeDetails
+    {
+    }
+
+    public enum RecurrenceFrequency
+    {
+        Daily,
+        Weekly,
+        Fortnightly,
+        Monthly,
+        Yearly
+    }
+
+    public enum DayOfTheWeek
+    {
+        Monday,
+        Tuesday,
+        Wednesday,
+        Thursday,
+        Friday,
+        Saturday,
+        Sunday
     }
 }
